requirements.txt
# Core Dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Robotics Libraries
pyserial>=3.5
pypot>=4.0.0
dynamixel-sdk>=3.7.0

# Computer Vision & AI
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
ultralytics>=8.0.0
onnx>=1.10.0
onnxruntime>=1.9.0

# LiDAR Processing
rplidar-roboticia>=0.9.2
python-pcl>=0.3.0
open3d>=0.15.0

# Navigation & SLAM
scikit-learn>=1.0.0
networkx>=2.6.0
pyyaml>=5.4.0

# Communication & Networking
paho-mqtt>=1.5.0
flask>=2.0.0
websockets>=10.0

# Utilities
click>=8.0.0
tqdm>=4.62.0
loguru>=0.5.0
psutil>=5.8.0

# Development & Testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7.0
flake8>=3.9.0

# Optional NPU Support (uncomment as needed)
# pycoral>=2.0.0  # For Google Coral Edge TPU
# jetson-stats>=4.0.0  # For NVIDIA Jetson devices 